+++ app/src/main/java/com/agile/ondemand/url/Url.java
@@ -5,7 +5,9 @@ import retrofit2.converter.gson.GsonConverterFactory;
 
 public class Url {
     // public static final String base_url = "http://192.168.1.11:3000/";
-    public static final String base_url = "http://10.1.18.43:3007/";
+ //   public static final String base_url = "http://10.1.18.43:3007/";
+    public static final String base_url = "http://192.168.100.247:3007/";
+
     //public static final String base_url = "http://10.0.2.2:3007/";
 //    public static final String base_url = "http://192.168.9.101:3012";
     // public static final String base_url = "http://172.100.100.5:3000/";
