--- app/src/test/java/com/agile/ondemand/testValidations.java
@@ -1,65 +1,0 @@
-package com.agile.ondemand;
-
-import com.agile.ondemand.bll.Validation;
-
-import org.junit.Test;
-
-import static org.junit.Assert.assertTrue;
-
-public class testValidations {
-
-    @Test
-    public void testEmailValidation(){
-
-
-    }
-    @Test
-    public void testUsernameValidation(){
-
-
-
-    }
-
-
-//        @Test
-//        public void emailValidator_CorrectEmailSimple_ReturnsTrue() {
-//            assertTrue(Validation.validateEmail("name@email.com"));
-//        }
-
-
-//
-//        @Test
-//        public void emailValidator_CorrectEmailSubDomain_ReturnsTrue() {
-//            assertTrue(EmailValidator.isValidEmail("name@email.co.np"));
-//        }
-//
-//        @Test
-//        public void emailValidator_InvalidEmailNoTld_ReturnsFalse() {
-//            assertFalse(EmailValidator.isValidEmail("name@email"));
-//        }
-//
-//        @Test
-//        public void emailValidator_InvalidEmailDoubleDot_ReturnsFalse() {
-//            assertFalse(EmailValidator.isValidEmail("name@email..com"));
-//        }
-//
-//        @Test
-//        public void emailValidator_InvalidEmailNoUsername_ReturnsFalse() {
-//            assertFalse(EmailValidator.isValidEmail("@email.com"));
-//        }
-//
-//        @Test
-//        public void emailValidator_EmptyString_ReturnsFalse() {
-//            assertFalse(EmailValidator.isValidEmail(""));
-//        }
-//
-//        @Test
-//        public void emailValidator_NullEmail_ReturnsFalse() {
-//            assertFalse(EmailValidator.isValidEmail(null));
-//        }
-//
-
-
-
-
-}
